refactor(dashboard): extract ws_agent_log block to agent_log.rs (#1266)
routes.rs: 2378 -> 2176
agent_log.rs: NEW 211

Moves WS_AGENT_LOG_ROUTE, sanitize_agent_name, agent_log_path,
ws_agent_log_handler, handle_agent_log_ws, wait_for_file.

Updates tests_routes_a and tests_routes_b to import from new module.
