+++ src/operator_commands_dashboard/mod.rs
@@ -12,6 +12,7 @@ mod chat;
 mod logs;
 mod registry;
 mod metrics;
+mod agent_log;
 pub(crate) mod routes;
 
 #[cfg(test)]
