+++ src/operator_commands_dashboard/tests_routes_b.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 mod tests_b {
     use crate::operator_commands_dashboard::routes::*;
+    use crate::operator_commands_dashboard::agent_log::{sanitize_agent_name, agent_log_path, WS_AGENT_LOG_ROUTE};
     use crate::operator_commands_dashboard::tmux::parse_tmux_sessions;
     use crate::operator_commands_dashboard::hosts::{host_entry_name, load_hosts};
     use crate::operator_commands_dashboard::memory::{build_agent_graph, classify_agent_layer};
